allow refering to $host:name.ip from any phase/order
Now provisioning is done with all hosts DOWN. Hosts are deployed in parallel
and some are waiting for others, to get their IP resolved. There is
currently no detection of deadlock (A refers to B and B refers to A). Then
hosts are started as usual with respecting order/phase